+++ examples/lib/stories/input/tappables.dart
@@ -17,7 +17,6 @@ class TappableSquare extends PositionComponent with Tappable {
 
   @override
   void render(Canvas canvas) {
-    super.render(canvas);
     canvas.drawRect(size.toRect(), _beenPressed ? _grey : _white);
   }