+++ packages/flame/lib/src/components/shape_component.dart
@@ -34,7 +34,6 @@ class ShapeComponent extends PositionComponent {
 
   @override
   void render(Canvas canvas) {
-    super.render(canvas);
     shape.render(canvas, paint);
   }